:sparkles: (upload): 添加文件操作方法并优化文件删除逻辑
添加 `getFile` 和 `getFileByPath` 方法用于根据 URI 名称和路径获取文件。优化 `removeFile` 和 `removeFiles` 方法，增加空值检查和日志记录，确保文件删除逻辑更加健壮。

+++ FileDao.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace nova\plugin\upload;
 
 use nova\framework\log\File;
+use nova\framework\log\Logger;
 use nova\plugin\orm\object\Dao;
 
 class FileDao extends Dao
@@ -15,8 +16,21 @@ class FileDao extends Dao
         }
         $this->update()->set(['is_temp'=>true])->where(['link_id' => $link_id])->commit();
         foreach ($uri_name as $name){
+            if (empty($name)) {
+                continue;
+            }
             $this->update()->set(['is_temp'=>false,'link_id'=>$link_id])->where(['uri_name' => $name])->commit();
         }
+    }
+
+    public function getFile($uri_name): ?FileModel
+    {
+        return $this->find(null, ['uri_name' => $uri_name]);
+    }
+
+    public function getFileByPath($path): ?FileModel
+    {
+        return $this->find(null, ['path' => $path]);
     }
 
     private function extraImage($content, $pattern): array
@@ -48,34 +62,54 @@ class FileDao extends Dao
 
     public function removeFile($uri_name): void
     {
+
+        if (empty($uri_name)) {
+            return;
+        }
+        /**
+         * @var FileModel $file
+         */
         $file  = $this->find(null,['uri_name' => $uri_name]);
 
-        if ($file && file_exists($file->path)) {
-            unlink($file->path);
-        }
+        if (empty($file->link_id)) return;
 
-        $this->delete()->where(['uri_name' => $uri_name])->commit();
+        if ($file && file_exists($file->path)) {
+            //Logger::info("删除文件(removeFile)".print_r($file,true).":".$file->path);
+            File::del($file->path);
+        }
+        $this->delete()->where(['id' => $file->id])->commit();
+        Logger::info(print_r(debug_backtrace(), true));
     }
 
 
     public function removeFiles($link_id): void
     {
+        if (empty($link_id)) {
+            return;
+        }
         $files = $this->getAll(null,['link_id' => $link_id]);
         /** @var FileModel $file */
         foreach ($files["data"] as $file) {
+            if (empty($file->uri_name)) {
+                continue;
+            }
             File::del($file->path);
+            //    Logger::info("删除文件(removeFiles)".print_r($file,true).":".$file->path);
         }
         $this->delete()->where(['link_id' => $link_id])->commit();
+        // Logger::info(print_r(debug_backtrace(),true));
     }
 
     public function removeTempFiles(): void
     {
-        $timeouts = time() - 3600 * 12;// 12 hour
+        $timeouts = time() - 3600 * 12;//- 3600 * 12;// 12 hour
         $files = $this->getAll(null,['is_temp' => true,"create_time < ".$timeouts]);
         /** @var FileModel $file */
         foreach ($files["data"] as $file) {
             File::del($file->path);
+            //    Logger::info("删除临时文件(removeTempFiles)".print_r($file,true).":".$file->path);
         }
-        $this->delete()->where(['is_temp' => true,"create_time < ".$timeouts])->commit();
+        //  Logger::info(print_r(debug_backtrace(),true));
+        $this->delete()->where(['is_temp' => true, "create_time < " . $timeouts])->commit();
     }
 }
